draw reactive time
draw reactive time by pinia
